basic 2 layer GNN class
