--- tests/test_gnn-model.py
@@ -1,2 +1,0 @@
-def test_forward():
-    assert False
